refactor: rename test/util/index to test/helper/index
